requirements.txt
python-telegram-bot
python-dotenv
google-generativeai
PyGithub
